Refactored DashboardAPI
